Use base system's python-cryptography
Without this patch, ansible install through pip tries to install
cryptography (as dep), which will fail because no rust runtime is
installed on those containers.

This fixes it by ensuring cryptography is already installed in the
OS, so it doesn't need to be reinstalled through pip.
This will only last as long as the cryptography installed from the
OS is not too low for ansible requirements. When this will be
the case, we can still decide to use constraints, or install
rust.